requirements.txt
Django==4.*
djangorestframework==3.13.1
PyYAML
uritemplate
markdown
Pygments
django-filter
django-guardian
drf-yasg==1.20.*
djangorestframework-camel-case==1.3.*
django-db-comments==0.2.2
coverage==6.4.*